+++ bee-message/tests/payload.rs
@@ -10,7 +10,7 @@ use bee_message::{
         milestone::{MilestoneEssence, MilestoneId, MilestonePayload},
         receipt::{MigratedFundsEntry, ReceiptPayload, TailTransactionHash},
         transaction::{RegularTransactionEssence, TransactionEssence, TransactionId, TransactionPayloadBuilder},
-        IndexationPayload, Payload, TreasuryTransactionPayload,
+        Payload, TaggedDataPayload, TreasuryTransactionPayload,
     },
     signature::{Ed25519Signature, Signature},
     unlock_block::{ReferenceUnlockBlock, SignatureUnlockBlock, UnlockBlock, UnlockBlocks},
@@ -99,14 +99,14 @@ fn milestone() {
 }
 
 #[test]
-fn indexation() {
-    let payload: Payload = IndexationPayload::new(rand_bytes(32), vec![]).unwrap().into();
+fn tagged_data() {
+    let payload: Payload = TaggedDataPayload::new(rand_bytes(32), vec![]).unwrap().into();
 
     let packed = payload.pack_to_vec();
 
-    assert_eq!(payload.kind(), 2);
+    assert_eq!(payload.kind(), 5);
     assert_eq!(payload.packed_len(), packed.len());
-    assert!(matches!(payload, Payload::Indexation(_)));
+    assert!(matches!(payload, Payload::TaggedData(_)));
 }
 
 #[test]