+++ bee-storage/bee-storage-sled/src/access/fetch.rs
@@ -12,7 +12,6 @@ use bee_message::{
     address::Ed25519Address,
     milestone::{Milestone, MilestoneIndex},
     output::OutputId,
-    payload::indexation::PaddedIndex,
     Message, MessageId,
 };
 use bee_storage::{access::Fetch, backend::StorageBackend, system::System};
@@ -68,25 +67,6 @@ impl Fetch<MessageId, Vec<MessageId>> for Storage {
                     Ok(MessageId::from(child))
                 })
                 .take(self.config.storage.fetch_edge_limit)
-                .collect::<Result<Vec<MessageId>, Self::Error>>()?,
-        ))
-    }
-}
-
-impl Fetch<PaddedIndex, Vec<MessageId>> for Storage {
-    fn fetch(&self, index: &PaddedIndex) -> Result<Option<Vec<MessageId>>, <Self as StorageBackend>::Error> {
-        Ok(Some(
-            self.inner
-                .open_tree(TREE_INDEX_TO_MESSAGE_ID)?
-                .scan_prefix(index)
-                .map(|result| {
-                    let (key, _) = result?;
-                    let (_, message_id) = key.split_at(PaddedIndex::LENGTH);
-                    // Unpacking from storage is fine.
-                    let message_id: [u8; MessageId::LENGTH] = message_id.try_into().unwrap();
-                    Ok(MessageId::from(message_id))
-                })
-                .take(self.config.storage.fetch_index_limit)
                 .collect::<Result<Vec<MessageId>, Self::Error>>()?,
         ))
     }